// apps/admin/src/api/generated/schemas/listAdminPaymentPoTransactionsParams.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * X-Ear CRM API
 * Auto-generated from Flask backend routes
 * OpenAPI spec version: 1.0.0
 */

export type ListAdminPaymentPoTransactionsParams = {
provider?: string | null;
start_date?: string | null;
end_date?: string | null;
/**
 * @minimum 1
 * @maximum 500
 */
limit?: number;
};