// apps/admin/src/api/generated/schemas/responseEnvelopeListPaymentRecordRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseEnvelopeListPaymentRecordReadData } from './responseEnvelopeListPaymentRecordReadData';
import type { ResponseEnvelopeListPaymentRecordReadMessage } from './responseEnvelopeListPaymentRecordReadMessage';
import type { ResponseEnvelopeListPaymentRecordReadError } from './responseEnvelopeListPaymentRecordReadError';
import type { ResponseEnvelopeListPaymentRecordReadMeta } from './responseEnvelopeListPaymentRecordReadMeta';
import type { ResponseEnvelopeListPaymentRecordReadRequestId } from './responseEnvelopeListPaymentRecordReadRequestId';

export interface ResponseEnvelopeListPaymentRecordRead {
  success?: boolean;
  data?: ResponseEnvelopeListPaymentRecordReadData;
  message?: ResponseEnvelopeListPaymentRecordReadMessage;
  error?: ResponseEnvelopeListPaymentRecordReadError;
  meta?: ResponseEnvelopeListPaymentRecordReadMeta;
  requestId?: ResponseEnvelopeListPaymentRecordReadRequestId;
  timestamp?: string;
}